Move EditorConfig and ConfigDefaults to lsystem-app-model
Extract the two application-model config families out of lsystem-core
into dedicated modules in lsystem-app-model:

- editor_config: EditorConfig, EditorGenerationConfig, EditorColorConfig,
  EditorLineColorConfig, ConfigSource, ConfigDocument, ParseConfigError,
  and all Raw* deserialization types with their TryFrom conversions and
  TOML manipulation helpers.
- config_defaults: ConfigDefaults, TurtleDefaults, ColorDefaults,
  LineColorDefaults, DefaultLineColorMode, GradientDefaults,
  HueCycleDefaults, and defaults.toml (moved from lsystem-core/src/).

lsystem-core now holds only runtime types (Config, GenerationConfig,
ColorConfig, LineColorConfig, Rgb, Dimensions, ConfigError) and drops
the serde and toml_edit dependencies. ConfigError loses its TOML
variants, which move to ParseConfigError in lsystem-app-model.

Also fixes the normalize_line_color_for_render boundary leak: resolved
Config.colors.line now faithfully reflects authored intent; callers that
allocate topological-depth geometry use the explicit conjunction
needs_topological_depth() && has_stack_directives() instead of relying
on silent rewriting at resolve time. SVG export tests are rewritten to
construct Config values directly without TOML parsing.
