+++ crates/lsystem-app-model/src/lib.rs
@@ -1,6 +1,8 @@
 pub(crate) mod animation;
 pub(crate) mod color;
+pub(crate) mod config_defaults;
 pub(crate) mod config_workspace;
+pub(crate) mod editor_config;
 pub(crate) mod presets;
 pub(crate) mod util;
 
@@ -11,8 +13,16 @@ pub use animation::{
 pub use color::{
     ColorControlMemory, LineColorMode, line_color_for_controls, selected_line_color_mode,
 };
+pub use config_defaults::{
+    ColorDefaults, ConfigDefaults, DefaultLineColorMode, GradientDefaults, HueCycleDefaults,
+    LineColorDefaults, ParseConfigError, TurtleDefaults,
+};
 pub use config_workspace::{
     CleanMut, ConfigEntry, ConfigWorkspace, ConfigWorkspaceError, DirtyMut, EntryViewMut,
+};
+pub use editor_config::{
+    ConfigDocument, ConfigSource, EditorColorConfig, EditorConfig, EditorGenerationConfig,
+    EditorLineColorConfig,
 };
 pub use presets::load_presets;
 pub use util::sanitize_filename;
