+++ crates/lsystem-web-app/src/presets.rs
@@ -1,4 +1,4 @@
-use lsystem_core::EditorGenerationConfig;
+use lsystem_app_model::EditorGenerationConfig;
 
 pub(crate) fn max_iterations_for_editor_config(generation: &EditorGenerationConfig) -> u32 {
     let max_seg = lsystem_renderer::line_renderer::max_segments_for_line_color(
